requirements.txt
fastapi==0.99.1
langchain==0.0.225
python-dotenv==1.0.0
uvicorn==0.22.0
openai==0.27.8